extract path utilities for use in other bootstrap 5 related management commands
