+++ corehq/apps/hqwebapp/management/commands/migrate_app_to_bootstrap5.py
@@ -4,7 +4,6 @@ from pathlib import Path
 from django.conf import settings
 from django.core.management import BaseCommand
 
-import corehq
 from corehq.apps.hqwebapp.utils.bootstrap import BOOTSTRAP_3, BOOTSTRAP_5
 from corehq.apps.hqwebapp.utils.bootstrap.changes import (
     get_spec,
@@ -18,8 +17,13 @@ from corehq.apps.hqwebapp.utils.bootstrap.changes import (
     file_contains_reference_to_path,
     replace_path_references,
 )
+from corehq.apps.hqwebapp.utils.bootstrap.paths import (
+    COREHQ_BASE_DIR,
+    get_app_template_folder,
+    get_app_static_folder,
+    get_short_path,
+)
 
-COREHQ_BASE_DIR = Path(corehq.__file__).resolve().parent
 COLOR_RED = "91"
 COLOR_GREEN = "92"
 COLOR_YELLOW = "93"
@@ -91,14 +95,6 @@ class Command(BaseCommand):
         self.stdout.write("Please review the next steps here:"
                           "\tcommcarehq.org/styleguide/b5/migration/#update-stylesheets\n\n\n")
 
-    @staticmethod
-    def _get_app_template_folder(app_name):
-        return COREHQ_BASE_DIR / "apps" / app_name / "templates" / app_name
-
-    @staticmethod
-    def _get_app_static_folder(app_name):
-        return COREHQ_BASE_DIR / "apps" / app_name / "static" / app_name
-
     def _get_files_for_migration(self, files, file_name, do_re_check):
         if file_name is not None:
             files = [t for t in files if file_name in str(t)]
@@ -115,18 +111,18 @@ class Command(BaseCommand):
         return files
 
     def get_templates_for_migration(self, app_name, template_name, do_re_check):
-        app_template_folder = self._get_app_template_folder(app_name)
+        app_template_folder = get_app_template_folder(app_name)
         app_templates = [f for f in app_template_folder.glob('**/*') if f.is_file()]
         return self._get_files_for_migration(app_templates, template_name, do_re_check)
 
     def get_js_files_for_migration(self, app_name, js_name, do_re_check):
-        app_static_folder = self._get_app_static_folder(app_name)
+        app_static_folder = get_app_static_folder(app_name)
         app_js_files = [f for f in app_static_folder.glob('**/*.js') if f.is_file()]
         return self._get_files_for_migration(app_js_files, js_name, do_re_check)
 
     def migrate_files(self, files, app_name, spec, is_template):
         for file_path in files:
-            short_path = self.get_short_path(app_name, file_path, is_template)
+            short_path = get_short_path(app_name, file_path, is_template)
             self.clear_screen()
             file_type = "templates" if is_template else "javascript"
             self.stdout.write(self.format_header(f"Migrating {app_name} {file_type}..."))
@@ -165,7 +161,7 @@ class Command(BaseCommand):
                 if line_changelog:
                     file_changelog.extend(line_changelog)
 
-            short_path = self.get_short_path(app_name, file_path, is_template)
+            short_path = get_short_path(app_name, file_path, is_template)
             if has_changes:
                 self.clear_screen()
                 self.stdout.write(
@@ -236,7 +232,7 @@ class Command(BaseCommand):
                           "in the Bootstrap 5 version of this file.\n\n")
 
     def record_file_changes(self, template_path, app_name, changelog, is_template):
-        short_path = self.get_short_path(app_name, template_path.parent, is_template)
+        short_path = get_short_path(app_name, template_path.parent, is_template)
         readme_directory = Path(settings.BOOTSTRAP_MIGRATION_LOGS_DIR) / short_path
         readme_directory.mkdir(parents=True, exist_ok=True)
         extension = '.html' if is_template else '.js'
@@ -253,7 +249,7 @@ class Command(BaseCommand):
         self.stdout.write("** Please make a note of this for reviewing later.\n\n\n")
 
     def save_re_checked_file_changes(self, app_name, file_path, changed_lines, is_template):
-        short_path = self.get_short_path(app_name, file_path, is_template)
+        short_path = get_short_path(app_name, file_path, is_template)
 
         confirm = self.get_confirmation(f"\nSave changes to {short_path}?")
 
@@ -267,7 +263,7 @@ class Command(BaseCommand):
         self.suggest_commit_message(f"re-ran migration for {short_path}")
 
     def split_files_and_refactor(self, app_name, file_path, bootstrap3_lines, bootstrap5_lines, is_template):
-        short_path = self.get_short_path(app_name, file_path, is_template)
+        short_path = get_short_path(app_name, file_path, is_template)
 
         confirm = self.get_confirmation(f'\nSplit {short_path} into Bootstrap 3 and Bootstrap 5 versions '
                                         f'and update references?')
@@ -276,8 +272,8 @@ class Command(BaseCommand):
             return
 
         bootstrap3_path, bootstrap5_path = self.get_split_file_paths(file_path)
-        bootstrap3_short_path = self.get_short_path(app_name, bootstrap3_path, is_template)
-        bootstrap5_short_path = self.get_short_path(app_name, bootstrap5_path, is_template)
+        bootstrap3_short_path = get_short_path(app_name, bootstrap3_path, is_template)
+        bootstrap5_short_path = get_short_path(app_name, bootstrap5_path, is_template)
         self.stdout.write(f"\n\nSplitting files:\n"
                           f"\n\t{bootstrap3_short_path}"
                           f"\n\t{bootstrap5_short_path}\n\n")
@@ -310,10 +306,10 @@ class Command(BaseCommand):
             file.writelines(bootstrap5_lines)
 
     def _get_migrated_files(self, app_name):
-        app_template_folder = self._get_app_template_folder(app_name)
+        app_template_folder = get_app_template_folder(app_name)
         migrated_files = [f for f in app_template_folder.glob('**/*')
                           if f.is_file() and '/bootstrap3/' in str(f) and '/crispy/' not in str(f)]
-        app_static_folder = self._get_app_static_folder(app_name)
+        app_static_folder = get_app_static_folder(app_name)
         migrated_files.extend(
             f for f in app_static_folder.glob('**/*.js')
             if f.is_file() and '/bootstrap3/' in str(f)
@@ -326,10 +322,10 @@ class Command(BaseCommand):
         self.stdout.write(f"\n\nVerifying that references to migrated files "
                           f"in {app_name} have been updated...")
         migrated_files = self._get_migrated_files(app_name)
-        template_path = self._get_app_template_folder(app_name)
+        template_path = get_app_template_folder(app_name)
         for file_path in migrated_files:
             is_template = file_path.is_relative_to(template_path)
-            new_reference = self.get_short_path(app_name, file_path, is_template)
+            new_reference = get_short_path(app_name, file_path, is_template)
             old_reference = new_reference.replace("/bootstrap3/", "/")
             references = self.update_and_get_references(
                 old_reference,
@@ -402,17 +398,6 @@ class Command(BaseCommand):
         bootstrap3_folder.mkdir(parents=True, exist_ok=True)
         bootstrap5_folder.mkdir(parents=True, exist_ok=True)
         return bootstrap3_folder / file_path.name, bootstrap5_folder / file_path.name
-
-    @staticmethod
-    def get_short_path(app_name, full_path, is_template):
-        if is_template:
-            replace_path = COREHQ_BASE_DIR / "apps" / app_name / "templates"
-        else:
-            replace_path = COREHQ_BASE_DIR / "apps" / app_name / "static"
-        return str(full_path).replace(
-            str(replace_path) + '/',
-            ''
-        )
 
     @staticmethod
     def select_option_from_prompt(prompt, options):
